feat: remake documents sync
